added lots of testing files
