+++ backend/api/rest/manifest/validator/tests/manifest2.json
@@ -1,0 +1,3 @@
+{
+  "project": {}
+}
