+++ backend/api/rest/tests/sample_project_manifest.json5
@@ -1,0 +1,51 @@
+{
+  "title": "test_project",
+  "url": "www.github.com/knex/",
+  "analysis": {
+    "url": "here-you-find-the-analysis.com",
+    "description": "thats what this analysis is about"
+  },
+  "authors": [
+    {
+      "name": "test_owner",
+      "email": "email@address.com"
+    }
+  ],
+  "team": {
+    "team_name": "teamname",
+    "members": [
+      {
+        "name": "member1",
+        "email": "member1@team.com"
+      },
+      {
+        "name": "member2",
+        "email": "member2@team.com"
+      }
+    ]
+  },
+  "status": "DONE",
+  "description": "can this be a multiline description?",
+  "date_creation": "1970-01-01",
+  "date_last_updated": "2017-05-11",
+  "tags": [
+    "tag",
+    "bio"
+  ],
+  "future_work": [
+    {
+      "title": "this is something related to this work",
+      "author": "the_author_of_this_awesome_work",
+      "email": "email@example.com",
+      "description": "some description of the work that will be done in the future"
+    }
+  ],
+  "related_projects": [
+    {
+      "id": 0,
+      "name": "name of the listed project",
+      "author": "author",
+      "url": "www.my.url"
+    }
+  ]
+}
